Major restructuring
* Split request into actions for responding and actions for obtaining
  the response.

-> IOFWDRequest now only deals with moving the response to the client
  using the correct protocol.

-> Created a factory mapping a request onto a task (i.e. something that
will complete the request)
Started implemening thread tasks

-> another iteration over the rescheduling mechanism

+++ src/iofwdutil/IOFWDLog.cpp
@@ -170,7 +170,7 @@ void IOFWDLog::setLogLevelOverride (const std::string & source)
       return; 
 
    /* remove global entries from list */
-   for (int i=loglevel_override_.size()-1; i>=0; --i)
+   for (int i=static_cast<int>(loglevel_override_.size())-1; i>=0; --i)
    {
       if (loglevel_override_[i].first.size()== 0)
       {
